# 🚀 Migrate CI/CD to ruff for 30-40x performance improvement