--- a/.github/workflows/ci.yml
+++ b/.github/workflows/ci.yml
@@ -40,22 +40,13 @@ jobs:
       run: |
         python -m pip install --upgrade pip
         pip install -r requirements.txt
-        pip install pytest pytest-cov black isort flake8 mypy
+        pip install pytest pytest-cov ruff mypy
 
-    - name: Lint with flake8
+    - name: Lint and format check with ruff
       run: |
-        # stop the build if there are Python syntax errors or undefined names
-        flake8 app --count --select=E9,F63,F7,F82 --show-source --statistics
-        # exit-zero treats all errors as warnings. The GitHub editor is 127 chars wide
-        flake8 app --count --exit-zero --max-complexity=10 --max-line-length=127 --statistics
-
-    - name: Format check with black
-      run: |
-        black --check app tests
-
-    - name: Import sort check with isort
-      run: |
-        isort --check-only app tests
+        # Check formatting and linting (combines black, isort, flake8)
+        ruff check app tests
+        ruff format --check app tests
 
     - name: Type check with mypy
       run: |
@@ -209,17 +200,17 @@ jobs:
     - name: Install formatting tools
       run: |
         python -m pip install --upgrade pip
-        pip install black isort
+        pip install ruff
 
     - name: Auto-format code
       run: |
-        black app tests
-        isort app tests
+        ruff check --fix app tests
+        ruff format app tests
 
     - name: Commit changes
       run: |
         git config --local user.email "action@github.com"
         git config --local user.name "GitHub Action"
-        git diff --quiet || (git add . && git commit -m "Auto-format code with black and isort")
+        git diff --quiet || (git add . && git commit -m "Auto-format code with ruff")
         git push
       continue-on-error: true