… app/utils/__pycache__/code_executor.cpython-312.pyc
… uv.lock