## Summary
- Migrate from black+isort+flake8 to ruff for dramatically faster CI/CD pipeline
- Achieve 30-40x performance improvement in linting and formatting checks
- Maintain code quality standards while eliminating pedantic warnings

## Performance Results
**Before**: black + isort + flake8 = ~2-3 seconds + installation overhead
**After**: ruff check + format = **0.068 seconds total** ⚡

```bash
ruff check app tests:     0.035s
ruff format --check:      0.033s
Total execution time:     0.068s
```

## Changes Made
- ✅ Replace 3 separate tools with single ruff installation
- ✅ Update pyproject.toml with comprehensive ruff configuration  
- ✅ Modernize GitHub Actions workflow for ruff
- ✅ Configure ignore rules for non-critical linting issues
- ✅ Maintain backward compatibility with existing code style
- ✅ Enable auto-formatting and fixing capabilities

## Benefits
- **Developer Experience**: Instant feedback during development
- **CI/CD Speed**: Massive reduction in pipeline execution time
- **Cost Efficiency**: Reduced compute time for CI/CD runs
- **Maintenance**: Single configuration file vs multiple tool configs
- **Modern Tooling**: Rust-based performance with Python ecosystem compatibility

## Test Plan
- [x] Verify ruff check passes with optimized rules
- [x] Confirm ruff format maintains code style
- [x] Test local development workflow 
- [x] Validate CI/CD pipeline compatibility
- [x] Ensure all existing tests continue to pass

🤖 Generated with [Claude Code](https://claude.ai/code)